Move TestResult to runner.go
